worker/requirements.txt
redis-py == 3.4
spleeter == 1.5
boto3 == 1.12
# see https://github.com/deezer/spleeter/wiki/1.-Installation#using-pip
ffmpeg